feat(delay): move stokes_I function to transform module
